Added metadata from nova for ironic nodes
This is obviously useful to associate metrics with projects and users.
The nova names are a convenience and mostly for faster triage by users.

Effect on cardinality

There is a certain risk for a bit more cardinality when users change
their baremetal machines more often. Considering that these users
literally book full baremetal machines the costs for additional
Prometheus compute and long term storage is economically negliable.

Changed label names

Consulting with @jobrs brought up inconsistencies in our naming schemes.
We hope the names are now better than before.
We dropped the label job as these should never be set by the source of
the metric and definitely not from a service discovery because we don’t know
for what the discovery will be used in the future.